Move in-memory store into tests
